[tps] skin, style 패키지 삭제/articlePage패키지 추가 및 API추가, componentWork관련 desking패키지로 이동

+++ moka-core/moka-core-tps/src/main/java/jmnet/moka/core/tps/mvc/articlepage/repository/ArticlePageHistRepository.java
@@ -8,6 +8,6 @@ import jmnet.moka.core.tps.mvc.articlepage.entity.ArticlePageHist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ArticlePageHistRepository extends JpaRepository<ArticlePageHist, Integer>, JpaSpecificationExecutor<ArticlePageHist> {
+public interface ArticlePageHistRepository extends JpaRepository<ArticlePageHist, Long>, JpaSpecificationExecutor<ArticlePageHist> {
 
 }
